refactor: refactor single-agent
