+++ safepo/single_agent/cpo.py
@@ -23,6 +23,7 @@ import sys
 import time
 from collections import deque
 from distutils.util import strtobool
+from typing import Callable
 
 import numpy as np
 import torch
@@ -30,12 +31,10 @@ import torch.nn as nn
 import torch.optim
 from rich.progress import track
 from torch.nn.utils.clip_grad import clip_grad_norm_
-from torch.optim.lr_scheduler import ConstantLR, LinearLR
 from torch.utils.data import DataLoader, TensorDataset
 
 from safepo.common.buffer import VectorizedOnPolicyBuffer
 from safepo.common.env import make_env
-from safepo.common.lagrange import PIDLagrangian as Lagrange
 from safepo.common.logger import EpochLogger
 from safepo.common.model import ActorVCritic
 
@@ -48,21 +47,21 @@ def parse_args():
         "--device",
         type=str,
         default="cpu",
-        help="if toggled, cuda will be enabled by default",
+        help="the device (cpu or cuda) to run the code",
     )
     parser.add_argument(
-        "--torch-threads", type=int, default=1, help="number of threads for torch"
+        "--torch-threads", type=int, default=4, help="number of threads for torch"
     )
     parser.add_argument(
         "--num-envs",
         type=int,
-        default=1,
+        default=10,
         help="the number of parallel game environments",
     )
     parser.add_argument(
         "--total-steps",
         type=int,
-        default=1024000,
+        default=10000000,
         help="total timesteps of the experiments",
     )
     parser.add_argument(
@@ -87,19 +86,19 @@ def parse_args():
     )
     # general algorithm parameters
     parser.add_argument(
-        "--steps_per_epoch",
+        "--steps-per-epoch",
         type=int,
-        default=2048,
+        default=20000,
         help="the number of steps to run in each environment per policy rollout",
     )
     parser.add_argument(
         "--update-iters",
         type=int,
-        default=40,
+        default=10,
         help="the max iteration to update the policy",
     )
     parser.add_argument(
-        "--batch-size", type=int, default=64, help="the number of mini-batches"
+        "--batch-size", type=int, default=128, help="the number of mini-batches"
     )
     parser.add_argument(
         "--entropy_coef", type=float, default=0.0, help="coefficient of the entropy"
@@ -107,7 +106,7 @@ def parse_args():
     parser.add_argument(
         "--target-kl",
         type=float,
-        default=0.02,
+        default=0.01,
         help="the target KL divergence threshold",
     )
     parser.add_argument(
@@ -154,31 +153,17 @@ def parse_args():
         help="toggles cost advantages standardization",
     )
     parser.add_argument(
-        "--actor-lr",
-        type=float,
-        default=3e-4,
-        help="the learning rate of the actor network",
-    )
-    parser.add_argument(
         "--critic-lr",
         type=float,
-        default=3e-4,
+        default=1e-3,
         help="the learning rate of the critic network",
-    )
-    parser.add_argument(
-        "--linear-lr-decay",
-        type=lambda x: bool(strtobool(x)),
-        default=True,
-        nargs="?",
-        const=True,
-        help="toggles learning rate annealing for policy and value networks",
     )
     # logger parameters
     parser.add_argument(
         "--log-dir",
         type=str,
         default="../runs",
-        help="directory to save agent logs (default: ../runs)",
+        help="directory to save agent logs",
     )
     parser.add_argument(
         "--write-terminal",
@@ -194,7 +179,34 @@ def parse_args():
     )
     # algorithm specific parameters
     parser.add_argument(
-        "--clip", type=float, default=0.2, help="the surrogate clipping coefficient"
+        "--fvp-sample-freq",
+        type=int,
+        default=1,
+        help="the sub-sampling rate of the observation",
+    )
+    parser.add_argument(
+        "--cg-damping",
+        type=float,
+        default=0.1,
+        help="the damping value for conjugate gradient",
+    )
+    parser.add_argument(
+        "--cg-iters",
+        type=int,
+        default=15,
+        help="the number of conjugate gradient iterations",
+    )
+    parser.add_argument(
+        "--backtrack-iters",
+        type=int,
+        default=15,
+        help="the number of backtracking line search iterations",
+    )
+    parser.add_argument(
+        "--backtrack-coef",
+        type=float,
+        default=0.8,
+        help="the coefficient for backtracking line search",
     )
     parser.add_argument(
         "--cost-limit",
@@ -202,54 +214,101 @@ def parse_args():
         default=25.0,
         help="the cost limit for the safety constraint",
     )
-    parser.add_argument(
-        "--lagrangian-multiplier-init",
-        type=float,
-        default=0.001,
-        help="the initial value of the lagrangian multiplier",
-    )
-    parser.add_argument(
-        "--pid-kp", type=float, default=0.1, help="the Kp of PID controller"
-    )
-    parser.add_argument(
-        "--pid-ki", type=float, default=0.01, help="the Ki of PID controller"
-    )
-    parser.add_argument(
-        "--pid-kd", type=float, default=0.0, help="the Kd of PID controller"
-    )
-    parser.add_argument(
-        "--pid-d-delay", type=int, default=10, help="the delay of derivative term"
-    )
-    parser.add_argument(
-        "--pid-delta-p-ema-alpha",
-        type=float,
-        default=0.95,
-        help="the exponential moving average alpha of the proportional term of the PID controller.",
-    )
-    parser.add_argument(
-        "--pid-delta-d-ema-alpha",
-        type=float,
-        default=0.95,
-        help="the exponential moving average alpha of the derivative term of the PID controller.",
-    )
-    parser.add_argument(
-        "--sum-norm",
-        type=lambda x: bool(strtobool(x)),
-        default=True,
-        help="toggles to normalize the sum of the cost",
-    )
-    parser.add_argument(
-        "--diff-norm",
-        type=lambda x: bool(strtobool(x)),
-        default=True,
-        help="toggles to normalize the derivate of the cost",
-    )
-    parser.add_argument(
-        "--penalty-max", type=float, default=100.0, help="the max penalty coefficient"
-    )
 
     args = parser.parse_args()
     return args
+
+
+def get_flat_params_from(model: torch.nn.Module) -> torch.Tensor:
+    flat_params = []
+    for _, param in model.named_parameters():
+        if param.requires_grad:
+            data = param.data
+            data = data.view(-1)  # flatten tensor
+            flat_params.append(data)
+    assert flat_params, "No gradients were found in model parameters."
+    return torch.cat(flat_params)
+
+
+def conjugate_gradients(
+    fisher_product: Callable[[torch.Tensor], torch.Tensor],
+    policy: ActorVCritic,
+    fvp_obs: torch.Tensor,
+    vector_b: torch.Tensor,
+    num_steps: int = 10,
+    residual_tol: float = 1e-10,
+    eps: float = 1e-6,
+) -> torch.Tensor:
+    vector_x = torch.zeros_like(vector_b)
+    vector_r = vector_b - fisher_product(vector_x, policy, fvp_obs)
+    vector_p = vector_r.clone()
+    rdotr = torch.dot(vector_r, vector_r)
+
+    for _ in range(num_steps):
+        vector_z = fisher_product(vector_p, policy, fvp_obs)
+        alpha = rdotr / (torch.dot(vector_p, vector_z) + eps)
+        vector_x += alpha * vector_p
+        vector_r -= alpha * vector_z
+        new_rdotr = torch.dot(vector_r, vector_r)
+        if torch.sqrt(new_rdotr) < residual_tol:
+            break
+        vector_mu = new_rdotr / (rdotr + eps)
+        vector_p = vector_r + vector_mu * vector_p
+        rdotr = new_rdotr
+    return vector_x
+
+
+def set_param_values_to_model(model: torch.nn.Module, vals: torch.Tensor) -> None:
+    assert isinstance(vals, torch.Tensor)
+    i: int = 0
+    for _, param in model.named_parameters():
+        if param.requires_grad:  # param has grad and, hence, must be set
+            orig_size = param.size()
+            size = np.prod(list(param.size()))
+            new_values = vals[i : int(i + size)]
+            # set new param values
+            new_values = new_values.view(orig_size)
+            param.data = new_values
+            i += int(size)  # increment array position
+    assert i == len(vals), f"Lengths do not match: {i} vs. {len(vals)}"
+
+
+def get_flat_gradients_from(model: torch.nn.Module) -> torch.Tensor:
+    grads = []
+    for _, param in model.named_parameters():
+        if param.requires_grad and param.grad is not None:
+            grad = param.grad
+            grads.append(grad.view(-1))  # flatten tensor and append
+    assert grads, "No gradients were found in model parameters."
+    return torch.cat(grads)
+
+
+def fvp(
+    params: torch.Tensor,
+    policy: ActorVCritic,
+    fvp_obs: torch.Tensor,
+) -> torch.Tensor:
+    policy.actor.zero_grad()
+    current_distribution = policy.actor(fvp_obs)
+    with torch.no_grad():
+        old_distribution = policy.actor(fvp_obs)
+    kl = torch.distributions.kl.kl_divergence(
+        old_distribution, current_distribution
+    ).mean()
+
+    grads = torch.autograd.grad(kl, tuple(policy.actor.parameters()), create_graph=True)
+    flat_grad_kl = torch.cat([grad.view(-1) for grad in grads])
+
+    kl_p = (flat_grad_kl * params).sum()
+    grads = torch.autograd.grad(
+        kl_p,
+        tuple(policy.actor.parameters()),
+        retain_graph=False,
+    )
+
+    flat_grad_grad_kl = torch.cat([grad.contiguous().view(-1) for grad in grads])
+
+    return flat_grad_grad_kl + params * args.cg_damping
 
 
 def main(args):
@@ -277,17 +336,6 @@ def main(args):
         obs_dim=obs_space.shape[0],
         act_dim=act_space.shape[0],
     ).to(device)
-    actor_optimizer = torch.optim.Adam(policy.actor.parameters(), lr=args.actor_lr)
-    if args.linear_lr_decay:
-        actor_scheduler = LinearLR(
-            actor_optimizer,
-            start_factor=1.0,
-            end_factor=0.0,
-            total_iters=epochs,
-            verbose=True,
-        )
-    else:
-        actor_scheduler = ConstantLR(actor_optimizer)
     reward_critic_optimizer = torch.optim.Adam(
         policy.reward_critic.parameters(), lr=args.critic_lr
     )
@@ -307,21 +355,6 @@ def main(args):
         standardized_adv_c=args.standardized_adv_c,
         device=device,
         num_envs=args.num_envs,
-    )
-
-    # setup lagrangian multiplier
-    lagrange = Lagrange(
-        pid_kd=args.pid_kd,
-        pid_kp=args.pid_kp,
-        pid_ki=args.pid_ki,
-        pid_d_delay=args.pid_d_delay,
-        pid_delta_p_ema_alpha=args.pid_delta_p_ema_alpha,
-        pid_delta_d_ema_alpha=args.pid_delta_d_ema_alpha,
-        sum_norm=args.sum_norm,
-        diff_norm=args.diff_norm,
-        penalty_max=args.penalty_max,
-        cost_limit=args.cost_limit,
-        lagrangian_multiplier_init=args.lagrangian_multiplier_init,
     )
 
     # set up the logger
@@ -467,40 +500,241 @@ def main(args):
 
         eval_end_time = time.time()
 
-        # update lagrange multiplier
-        ep_costs = logger.get_stats("Metrics/EpCost")
-        lagrange.update_lagrange_multiplier(ep_costs)
-
         # update policy
         data = buffer.get()
+        fvp_obs = data["obs"][:: args.fvp_sample_freq]
+        theta_old = get_flat_params_from(policy.actor)
+        policy.actor.zero_grad()
+        # compute loss_pi
+        temp_distribution = policy.actor(data["obs"])
+        log_prob = temp_distribution.log_prob(data["act"]).sum(dim=-1)
+        ratio = torch.exp(log_prob - data["log_prob"])
+        loss_pi_r = -(ratio * data["adv_r"]).mean()
+        loss_reward_before = loss_pi_r.item()
         old_distribution = policy.actor(data["obs"])
 
-        # comnpute advantage
-        advantage = data["adv_r"] - lagrange.lagrangian_multiplier * data["adv_c"]
-        advantage /= lagrange.lagrangian_multiplier + 1
+        loss_pi_r.backward()
+
+        grads = -get_flat_gradients_from(policy.actor)
+        x = conjugate_gradients(fvp, policy, fvp_obs, grads, args.cg_iters)
+        assert torch.isfinite(x).all(), "x is not finite"
+        xHx = torch.dot(x, fvp(x, policy, fvp_obs))
+        assert xHx.item() >= 0, "xHx is negative"
+        alpha = torch.sqrt(2 * args.target_kl / (xHx + 1e-8))
+
+        policy.actor.zero_grad()
+        temp_distribution = policy.actor(data["obs"])
+        log_prob = temp_distribution.log_prob(data["act"]).sum(dim=-1)
+        ratio = torch.exp(log_prob - data["log_prob"])
+        loss_pi_c = (ratio * data["adv_c"]).mean()
+        loss_cost_before = loss_pi_c.item()
+
+        loss_pi_c.backward()
+
+        b_grads = get_flat_gradients_from(policy.actor)
+        ep_costs = logger.get_stats("Metrics/EpCost") - args.cost_limit
+
+        p = conjugate_gradients(fvp, policy, fvp_obs, b_grads, args.cg_iters)
+        q = xHx
+        r = grads.dot(p)
+        s = b_grads.dot(p)
+
+        if b_grads.dot(b_grads) <= 1e-6 and ep_costs < 0:
+            # feasible step and cost grad is zero: use plain TRPO update...
+            A = torch.zeros(1)
+            B = torch.zeros(1)
+            optim_case = 4
+        else:
+            assert torch.isfinite(r).all(), "r is not finite"
+            assert torch.isfinite(s).all(), "s is not finite"
+
+            A = q - r**2 / (s + 1e-8)
+            B = 2 * args.target_kl - ep_costs**2 / (s + 1e-8)
+
+            if ep_costs < 0 and B < 0:
+                # point in trust region is feasible and safety boundary doesn't intersect
+                # ==> entire trust region is feasible
+                optim_case = 3
+            elif ep_costs < 0 <= B:
+                # point in trust region is feasible but safety boundary intersects
+                # ==> only part of trust region is feasible
+                optim_case = 2
+            elif ep_costs >= 0 and B >= 0:
+                # point in trust region is infeasible and cost boundary doesn't intersect
+                # ==> entire trust region is infeasible
+                optim_case = 1
+                logger.log("Alert! Attempting feasible recovery!", "yellow")
+            else:
+                # x = 0 infeasible, and safety half space is outside trust region
+                # ==> whole trust region is infeasible, try to fail gracefully
+                optim_case = 0
+                logger.log("Alert! Attempting infeasible recovery!", "red")
+
+        if optim_case in (3, 4):
+            # under 3 and 4 cases directly use TRPO method
+            alpha = torch.sqrt(2 * args.target_kl / (xHx + 1e-8))
+            nu_star = torch.zeros(1)
+            lambda_star = 1 / (alpha + 1e-8)
+            step_direction = alpha * x
+
+        elif optim_case in (1, 2):
+
+            def project(
+                data: torch.Tensor, low: torch.Tensor, high: torch.Tensor
+            ) -> torch.Tensor:
+                """Project data to [low, high] interval."""
+                return torch.clamp(data, low, high)
+
+            #  analytical Solution to LQCLP, employ lambda,nu to compute final solution of OLOLQC
+            #  λ=argmax(f_a(λ),f_b(λ)) = λa_star or λb_star
+            #  computing formula shown in appendix, lambda_a and lambda_b
+            lambda_a = torch.sqrt(A / B)
+            lambda_b = torch.sqrt(q / (2 * args.target_kl))
+            # λa_star = Proj(lambda_a ,0 ~ r/c)  λb_star=Proj(lambda_b,r/c~ +inf)
+            # where projection(str,b,c)=max(b,min(str,c))
+            # may be regarded as a projection from effective region towards safety region
+            r_num = r.item()
+            eps_cost = ep_costs + 1e-8
+            if ep_costs < 0:
+                lambda_a_star = project(
+                    lambda_a, torch.as_tensor(0.0), r_num / eps_cost
+                )
+                lambda_b_star = project(
+                    lambda_b, r_num / eps_cost, torch.as_tensor(torch.inf)
+                )
+            else:
+                lambda_a_star = project(
+                    lambda_a, r_num / eps_cost, torch.as_tensor(torch.inf)
+                )
+                lambda_b_star = project(
+                    lambda_b, torch.as_tensor(0.0), r_num / eps_cost
+                )
+
+            def f_a(lam: torch.Tensor) -> torch.Tensor:
+                return -0.5 * (A / (lam + 1e-8) + B * lam) - r * ep_costs / (s + 1e-8)
+
+            def f_b(lam: torch.Tensor) -> torch.Tensor:
+                return -0.5 * (q / (lam + 1e-8) + 2 * args.target_kl * lam)
+
+            lambda_star = (
+                lambda_a_star
+                if f_a(lambda_a_star) >= f_b(lambda_b_star)
+                else lambda_b_star
+            )
+
+            # discard all negative values with torch.clamp(x, min=0)
+            # Nu_star = (lambda_star * - r)/s
+            nu_star = torch.clamp(lambda_star * ep_costs - r, min=0) / (s + 1e-8)
+            # final x_star as final direction played as policy's loss to backward and update
+            step_direction = 1.0 / (lambda_star + 1e-8) * (x - nu_star * p)
+
+        else:  # case == 0
+            # purely decrease costs
+            # without further check
+            lambda_star = torch.zeros(1)
+            nu_star = torch.sqrt(2 * args.target_kl / (s + 1e-8))
+            step_direction = -nu_star * p
+
+        # get distance each time theta goes towards certain direction
+        step_frac = 1.0
+        # get and flatten parameters from pi-net
+        theta_old = get_flat_params_from(policy.actor)
+        # reward improvement, g-flat as gradient of reward
+        expected_reward_improve = grads.dot(step_direction)
+
+        kl = torch.zeros(1)
+        # while not within_trust_region and not finish all steps:
+        for step in range(args.backtrack_iters):
+            # get new theta
+            new_theta = theta_old + step_frac * step_direction
+            # set new theta as new actor parameters
+            set_param_values_to_model(policy.actor, new_theta)
+            # the last acceptance steps to next step
+            acceptance_step = step + 1
+
+            with torch.no_grad():
+                try:
+                    temp_distribution = policy.actor(data["obs"])
+                    log_prob = temp_distribution.log_prob(data["act"]).sum(dim=-1)
+                    ratio = torch.exp(log_prob - data["log_prob"])
+                    loss_reward = -(ratio * data["adv_r"]).mean()
+                except ValueError:
+                    step_frac *= args.backtrack_coef
+                    continue
+                # loss of cost of policy cost from real/expected reward
+                temp_distribution = policy.actor(data["obs"])
+                log_prob = temp_distribution.log_prob(data["act"]).sum(dim=-1)
+                ratio = torch.exp(log_prob - data["log_prob"])
+                loss_cost = (ratio * data["adv_c"]).mean()
+                # compute KL distance between new and old policy
+                current_distribution = policy.actor(data["obs"])
+                kl = torch.distributions.kl.kl_divergence(
+                    old_distribution, current_distribution
+                ).mean()
+            # compute improvement of reward
+            loss_reward_improve = loss_reward_before - loss_reward.item()
+            # compute difference of cost
+            loss_cost_diff = loss_cost.item() - loss_cost_before
+
+            logger.log(
+                f"Expected Improvement: {expected_reward_improve} Actual: {loss_reward_improve}",
+            )
+            # check whether there are nan.
+            if not torch.isfinite(loss_reward) and not torch.isfinite(loss_cost):
+                logger.log("WARNING: loss_pi not finite")
+            if not torch.isfinite(kl):
+                logger.log("WARNING: KL not finite")
+                continue
+            if loss_reward_improve < 0 if optim_case > 1 else False:
+                logger.log("INFO: did not improve improve <0")
+            # change of cost's range
+            elif loss_cost_diff > max(-ep_costs, 0):
+                logger.log(f"INFO: no improve {loss_cost_diff} > {max(-ep_costs, 0)}")
+            # check KL-distance to avoid too far gap
+            elif kl > args.target_kl:
+                logger.log(f"INFO: violated KL constraint {kl} at step {step + 1}.")
+            else:
+                # step only if surrogate is improved and we are
+                # within the trust region
+                logger.log(f"Accept step at i={step + 1}")
+                break
+            step_frac *= args.backtrack_coef
+        else:
+            # if didn't find a step satisfy those conditions
+            logger.log("INFO: no suitable step found...")
+            step_direction = torch.zeros_like(step_direction)
+            acceptance_step = 0
+
+        theta_new = theta_old + step_frac * step_direction
+        set_param_values_to_model(policy.actor, theta_new)
+
+        logger.store(
+            **{
+                "Misc/Alpha": alpha.item(),
+                "Misc/FinalStepNorm": torch.norm(step_direction).mean().item(),
+                "Misc/xHx": xHx.item(),
+                "Misc/gradient_norm": torch.norm(grads).mean().item(),
+                "Misc/H_inv_g": x.norm().item(),
+                "Misc/AcceptanceStep": acceptance_step,
+                "Loss/Loss_actor": (loss_pi_r + loss_pi_c).mean().item(),
+                "Train/KL": kl,
+            },
+        )
 
         dataloader = DataLoader(
             dataset=TensorDataset(
                 data["obs"],
-                data["act"],
-                data["log_prob"],
                 data["target_value_r"],
                 data["target_value_c"],
-                advantage,
             ),
             batch_size=args.batch_size,
             shuffle=True,
         )
-        update_counts = 0
-        final_kl = torch.ones_like(old_distribution.loc)
-        for i in track(range(args.update_iters), description="Updating..."):
+        for _ in track(range(args.update_iters), description="Updating..."):
             for (
                 obs_b,
-                act_b,
-                log_prob_b,
                 target_value_r_b,
                 target_value_c_b,
-                adv_b,
             ) in dataloader:
                 reward_critic_optimizer.zero_grad()
                 loss_r = nn.functional.mse_loss(
@@ -528,41 +762,13 @@ def main(args):
                 )
                 cost_critic_optimizer.step()
 
-                distribution = policy.actor(obs_b)
-                log_prob = distribution.log_prob(act_b).sum(dim=-1)
-                ratio = torch.exp(log_prob - log_prob_b)
-                ratio_cliped = torch.clamp(
-                    ratio,
-                    1 - args.clip,
-                    1 + args.clip,
-                )
-                loss_pi = -torch.min(ratio * adv_b, ratio_cliped * adv_b).mean()
-                actor_optimizer.zero_grad()
-                loss_pi.backward()
-                clip_grad_norm_(policy.actor.parameters(), args.max_grad_norm)
-                actor_optimizer.step()
-
                 logger.store(
                     **{
                         "Loss/Loss_reward_critic": loss_r.mean().item(),
                         "Loss/Loss_cost_critic": loss_c.mean().item(),
-                        "Loss/Loss_actor": loss_pi.mean().item(),
                     }
                 )
-            new_distribution = policy.actor(data["obs"])
-            kl = (
-                torch.distributions.kl.kl_divergence(old_distribution, new_distribution)
-                .sum(-1, keepdim=True)
-                .mean()
-                .item()
-            )
-            final_kl = kl
-            update_counts += 1
-            if kl > args.target_kl:
-                logger.log(f"Early stopping at iter {i + 1} due to reaching max kl")
-                break
         update_end_time = time.time()
-        actor_scheduler.step()
 
         # log data
         logger.log_tabular("Metrics/EpRet")
@@ -574,10 +780,7 @@ def main(args):
             logger.log_tabular("Metrics/EvalEpLen")
         logger.log_tabular("Train/Epoch", epoch + 1)
         logger.log_tabular("Train/TotalSteps", (epoch + 1) * args.steps_per_epoch)
-        logger.log_tabular("Train/StopIter", update_counts)
-        logger.log_tabular("Train/KL", final_kl)
-        logger.log_tabular("Train/LagragianMultiplier", lagrange.lagrangian_multiplier)
-        logger.log_tabular("Train/LR", actor_scheduler.get_last_lr()[0])
+        logger.log_tabular("Train/KL")
         logger.log_tabular("Loss/Loss_reward_critic")
         logger.log_tabular("Loss/Loss_cost_critic")
         logger.log_tabular("Loss/Loss_actor")
@@ -588,6 +791,12 @@ def main(args):
         logger.log_tabular("Time/Total", update_end_time - rollout_start_time)
         logger.log_tabular("Value/RewardAdv", data["adv_r"].mean().item())
         logger.log_tabular("Value/CostAdv", data["adv_c"].mean().item())
+        logger.log_tabular("Misc/Alpha")
+        logger.log_tabular("Misc/FinalStepNorm")
+        logger.log_tabular("Misc/xHx")
+        logger.log_tabular("Misc/gradient_norm")
+        logger.log_tabular("Misc/H_inv_g")
+        logger.log_tabular("Misc/AcceptanceStep")
 
         logger.dump_tabular()
         if epoch % 100 == 0:
