+++ safepo/single_agent/focops.py
@@ -23,14 +23,15 @@ import sys
 import time
 from collections import deque
 from distutils.util import strtobool
-from typing import Callable
 
 import numpy as np
 import torch
 import torch.nn as nn
 import torch.optim
 from rich.progress import track
+from torch.distributions import Normal
 from torch.nn.utils.clip_grad import clip_grad_norm_
+from torch.optim.lr_scheduler import ConstantLR, LinearLR
 from torch.utils.data import DataLoader, TensorDataset
 
 from safepo.common.buffer import VectorizedOnPolicyBuffer
@@ -51,18 +52,18 @@ def parse_args():
         help="if toggled, cuda will be enabled by default",
     )
     parser.add_argument(
-        "--torch-threads", type=int, default=1, help="number of threads for torch"
+        "--torch-threads", type=int, default=4, help="number of threads for torch"
     )
     parser.add_argument(
         "--num-envs",
         type=int,
-        default=1,
+        default=10,
         help="the number of parallel game environments",
     )
     parser.add_argument(
         "--total-steps",
         type=int,
-        default=1024000,
+        default=10000000,
         help="total timesteps of the experiments",
     )
     parser.add_argument(
@@ -89,17 +90,17 @@ def parse_args():
     parser.add_argument(
         "--steps_per_epoch",
         type=int,
-        default=2048,
+        default=20000,
         help="the number of steps to run in each environment per policy rollout",
     )
     parser.add_argument(
         "--update-iters",
         type=int,
-        default=10,
+        default=40,
         help="the max iteration to update the policy",
     )
     parser.add_argument(
-        "--batch-size", type=int, default=128, help="the number of mini-batches"
+        "--batch-size", type=int, default=64, help="the number of mini-batches"
     )
     parser.add_argument(
         "--entropy_coef", type=float, default=0.0, help="coefficient of the entropy"
@@ -107,7 +108,7 @@ def parse_args():
     parser.add_argument(
         "--target-kl",
         type=float,
-        default=0.01,
+        default=0.02,
         help="the target KL divergence threshold",
     )
     parser.add_argument(
@@ -154,10 +155,24 @@ def parse_args():
         help="toggles cost advantages standardization",
     )
     parser.add_argument(
+        "--actor-lr",
+        type=float,
+        default=3e-4,
+        help="the learning rate of the actor network",
+    )
+    parser.add_argument(
         "--critic-lr",
         type=float,
-        default=1e-3,
+        default=3e-4,
         help="the learning rate of the critic network",
+    )
+    parser.add_argument(
+        "--linear-lr-decay",
+        type=lambda x: bool(strtobool(x)),
+        default=True,
+        nargs="?",
+        const=True,
+        help="toggles learning rate annealing for policy and value networks",
     )
     # logger parameters
     parser.add_argument(
@@ -180,22 +195,7 @@ def parse_args():
     )
     # algorithm specific parameters
     parser.add_argument(
-        "--fvp-sample-freq",
-        type=int,
-        default=1,
-        help="the sub-sampling rate of the observation",
-    )
-    parser.add_argument(
-        "--cg-damping",
-        type=float,
-        default=0.1,
-        help="the damping value for conjugate gradient",
-    )
-    parser.add_argument(
-        "--cg-iters",
-        type=int,
-        default=15,
-        help="the number of conjugate gradient iterations",
+        "--clip", type=float, default=0.2, help="the surrogate clipping coefficient"
     )
     parser.add_argument(
         "--cost-limit",
@@ -215,101 +215,24 @@ def parse_args():
         default=0.035,
         help="the learning rate of the lagrangian multiplier",
     )
+    parser.add_argument(
+        "--lagrangian-upper-bound",
+        type=float,
+        default=2.0,
+        help="the upper bound of lagrange multiplier",
+    )
+    parser.add_argument(
+        "--focops-eta", type=float, default=0.02, help="the eta of the focops"
+    )
+    parser.add_argument(
+        "--focops-lam",
+        type=float,
+        default=1.5,
+        help="the hyperparameters related to the greediness of the algorithm",
+    )
 
     args = parser.parse_args()
     return args
-
-
-def get_flat_params_from(model: torch.nn.Module) -> torch.Tensor:
-    flat_params = []
-    for _, param in model.named_parameters():
-        if param.requires_grad:
-            data = param.data
-            data = data.view(-1)  # flatten tensor
-            flat_params.append(data)
-    assert flat_params, "No gradients were found in model parameters."
-    return torch.cat(flat_params)
-
-
-def conjugate_gradients(
-    fisher_product: Callable[[torch.Tensor], torch.Tensor],
-    policy: ActorVCritic,
-    fvp_obs: torch.Tensor,
-    vector_b: torch.Tensor,
-    num_steps: int = 10,
-    residual_tol: float = 1e-10,
-    eps: float = 1e-6,
-) -> torch.Tensor:
-    vector_x = torch.zeros_like(vector_b)
-    vector_r = vector_b - fisher_product(vector_x, policy, fvp_obs)
-    vector_p = vector_r.clone()
-    rdotr = torch.dot(vector_r, vector_r)
-
-    for _ in range(num_steps):
-        vector_z = fisher_product(vector_p, policy, fvp_obs)
-        alpha = rdotr / (torch.dot(vector_p, vector_z) + eps)
-        vector_x += alpha * vector_p
-        vector_r -= alpha * vector_z
-        new_rdotr = torch.dot(vector_r, vector_r)
-        if torch.sqrt(new_rdotr) < residual_tol:
-            break
-        vector_mu = new_rdotr / (rdotr + eps)
-        vector_p = vector_r + vector_mu * vector_p
-        rdotr = new_rdotr
-    return vector_x
-
-
-def set_param_values_to_model(model: torch.nn.Module, vals: torch.Tensor) -> None:
-    assert isinstance(vals, torch.Tensor)
-    i: int = 0
-    for _, param in model.named_parameters():
-        if param.requires_grad:  # param has grad and, hence, must be set
-            orig_size = param.size()
-            size = np.prod(list(param.size()))
-            new_values = vals[i : int(i + size)]
-            # set new param values
-            new_values = new_values.view(orig_size)
-            param.data = new_values
-            i += int(size)  # increment array position
-    assert i == len(vals), f"Lengths do not match: {i} vs. {len(vals)}"
-
-
-def get_flat_gradients_from(model: torch.nn.Module) -> torch.Tensor:
-    grads = []
-    for _, param in model.named_parameters():
-        if param.requires_grad and param.grad is not None:
-            grad = param.grad
-            grads.append(grad.view(-1))  # flatten tensor and append
-    assert grads, "No gradients were found in model parameters."
-    return torch.cat(grads)
-
-
-def fvp(
-    params: torch.Tensor,
-    policy: ActorVCritic,
-    fvp_obs: torch.Tensor,
-) -> torch.Tensor:
-    policy.actor.zero_grad()
-    current_distribution = policy.actor(fvp_obs)
-    with torch.no_grad():
-        old_distribution = policy.actor(fvp_obs)
-    kl = torch.distributions.kl.kl_divergence(
-        old_distribution, current_distribution
-    ).mean()
-
-    grads = torch.autograd.grad(kl, tuple(policy.actor.parameters()), create_graph=True)
-    flat_grad_kl = torch.cat([grad.view(-1) for grad in grads])
-
-    kl_p = (flat_grad_kl * params).sum()
-    grads = torch.autograd.grad(
-        kl_p,
-        tuple(policy.actor.parameters()),
-        retain_graph=False,
-    )
-
-    flat_grad_grad_kl = torch.cat([grad.contiguous().view(-1) for grad in grads])
-
-    return flat_grad_grad_kl + params * args.cg_damping
 
 
 def main(args):
@@ -337,6 +260,17 @@ def main(args):
         obs_dim=obs_space.shape[0],
         act_dim=act_space.shape[0],
     ).to(device)
+    actor_optimizer = torch.optim.Adam(policy.actor.parameters(), lr=args.actor_lr)
+    if args.linear_lr_decay:
+        actor_scheduler = LinearLR(
+            actor_optimizer,
+            start_factor=1.0,
+            end_factor=0.0,
+            total_iters=epochs,
+            verbose=True,
+        )
+    else:
+        actor_scheduler = ConstantLR(actor_optimizer)
     reward_critic_optimizer = torch.optim.Adam(
         policy.reward_critic.parameters(), lr=args.critic_lr
     )
@@ -381,6 +315,8 @@ def main(args):
     logger.save_config(dict_args)
     logger.setup_torch_saver(policy.actor)
     logger.log("Start with training.")
+
+    time.time()
 
     # training loop
     for epoch in range(epochs):
@@ -512,67 +448,41 @@ def main(args):
 
         # update policy
         data = buffer.get()
-        fvp_obs = data["obs"][:: args.fvp_sample_freq]
-        theta_old = get_flat_params_from(policy.actor)
-        policy.actor.zero_grad()
+        with torch.no_grad():
+            old_distribution = policy.actor(data["obs"])
+            old_mean = old_distribution.mean
+            old_std = old_distribution.stddev
 
         # comnpute advantage
         advantage = data["adv_r"] - lagrange.lagrangian_multiplier * data["adv_c"]
         advantage /= lagrange.lagrangian_multiplier + 1
 
-        # compute loss_pi
-        distribution = policy.actor(data["obs"])
-        log_prob = distribution.log_prob(data["act"]).sum(dim=-1)
-        ratio = torch.exp(log_prob - data["log_prob"])
-        loss_pi = -(ratio * advantage).mean()
-
-        loss_pi.backward()
-
-        grads = -get_flat_gradients_from(policy.actor)
-        x = conjugate_gradients(fvp, policy, fvp_obs, grads, args.cg_iters)
-        assert torch.isfinite(x).all(), "x is not finite"
-        xHx = torch.dot(x, fvp(x, policy, fvp_obs))
-        assert xHx.item() >= 0, "xHx is negative"
-        alpha = torch.sqrt(2 * args.target_kl / (xHx + 1e-8))
-        step_direction = x * alpha
-        assert torch.isfinite(step_direction).all(), "step_direction is not finite"
-
-        theta_new = theta_old + step_direction
-        set_param_values_to_model(policy.actor, theta_new)
-        with torch.no_grad():
-            new_distribution = policy.actor(data["obs"])
-            final_kl = (
-                torch.distributions.kl.kl_divergence(distribution, new_distribution)
-                .mean()
-                .item()
-            )
-
-        logger.store(
-            **{
-                "Misc/Alpha": alpha.item(),
-                "Misc/FinalStepNorm": torch.norm(step_direction).mean().item(),
-                "Misc/xHx": xHx.item(),
-                "Misc/gradient_norm": torch.norm(grads).mean().item(),
-                "Misc/H_inv_g": x.norm().item(),
-                "Loss/Loss_actor": loss_pi.mean().item(),
-                "Train/KL": final_kl,
-            },
-        )
-
         dataloader = DataLoader(
             dataset=TensorDataset(
                 data["obs"],
+                data["act"],
+                data["log_prob"],
                 data["target_value_r"],
                 data["target_value_c"],
+                advantage,
+                old_mean,
+                old_std,
             ),
             batch_size=args.batch_size,
             shuffle=True,
         )
-        for _ in track(range(args.update_iters), description="Updating..."):
+        update_counts = 0
+        final_kl = torch.ones_like(old_distribution.loc)
+        for i in track(range(args.update_iters), description="Updating..."):
             for (
                 obs_b,
+                act_b,
+                log_prob_b,
                 target_value_r_b,
                 target_value_c_b,
+                adv_b,
+                old_mean_b,
+                old_std_b,
             ) in dataloader:
                 reward_critic_optimizer.zero_grad()
                 loss_r = nn.functional.mse_loss(
@@ -600,13 +510,48 @@ def main(args):
                 )
                 cost_critic_optimizer.step()
 
+                old_distribution_b = Normal(loc=old_mean_b, scale=old_std_b)
+
+                distribution = policy.actor(obs_b)
+                log_prob = distribution.log_prob(act_b).sum(dim=-1)
+                ratio = torch.exp(log_prob - log_prob_b)
+                temp_kl = torch.distributions.kl_divergence(
+                    distribution, old_distribution_b
+                ).sum(-1, keepdim=True)
+
+                loss_pi = (temp_kl - (1 / args.focops_lam) * ratio * adv_b) * (
+                    temp_kl.detach() <= args.focops_eta
+                ).type(torch.float32)
+
+                loss_pi = loss_pi.mean()
+                loss_pi -= args.entropy_coef * distribution.entropy().mean()
+
+                actor_optimizer.zero_grad()
+                loss_pi.backward()
+                clip_grad_norm_(policy.actor.parameters(), args.max_grad_norm)
+                actor_optimizer.step()
+
                 logger.store(
                     **{
                         "Loss/Loss_reward_critic": loss_r.mean().item(),
                         "Loss/Loss_cost_critic": loss_c.mean().item(),
+                        "Loss/Loss_actor": loss_pi.mean().item(),
                     }
                 )
+            new_distribution = policy.actor(data["obs"])
+            kl = (
+                torch.distributions.kl.kl_divergence(old_distribution, new_distribution)
+                .sum(-1, keepdim=True)
+                .mean()
+                .item()
+            )
+            final_kl = kl
+            update_counts += 1
+            if kl > args.target_kl:
+                logger.log(f"Early stopping at iter {i + 1} due to reaching max kl")
+                break
         update_end_time = time.time()
+        actor_scheduler.step()
 
         # log data
         logger.log_tabular("Metrics/EpRet")
@@ -618,8 +563,10 @@ def main(args):
             logger.log_tabular("Metrics/EvalEpLen")
         logger.log_tabular("Train/Epoch", epoch + 1)
         logger.log_tabular("Train/TotalSteps", (epoch + 1) * args.steps_per_epoch)
-        logger.log_tabular("Train/KL")
+        logger.log_tabular("Train/StopIter", update_counts)
+        logger.log_tabular("Train/KL", final_kl)
         logger.log_tabular("Train/LagragianMultiplier", lagrange.lagrangian_multiplier)
+        logger.log_tabular("Train/LR", actor_scheduler.get_last_lr()[0])
         logger.log_tabular("Loss/Loss_reward_critic")
         logger.log_tabular("Loss/Loss_cost_critic")
         logger.log_tabular("Loss/Loss_actor")
@@ -630,11 +577,6 @@ def main(args):
         logger.log_tabular("Time/Total", update_end_time - rollout_start_time)
         logger.log_tabular("Value/RewardAdv", data["adv_r"].mean().item())
         logger.log_tabular("Value/CostAdv", data["adv_c"].mean().item())
-        logger.log_tabular("Misc/Alpha")
-        logger.log_tabular("Misc/FinalStepNorm")
-        logger.log_tabular("Misc/xHx")
-        logger.log_tabular("Misc/gradient_norm")
-        logger.log_tabular("Misc/H_inv_g")
 
         logger.dump_tabular()
         if epoch % 100 == 0:
